Split form unified repo
In an effort to keep Vivado commits (HW commits) apart form SDK or Vitis commits (SW commits), starting Vivado 2020.1 there will be separate repositories for HW and SW. This is the first commit which ports the SW workspace from the branch of the unified repository corresponding HW repository.

The hardware handoff in the for of *.xsa has been intentionally placed in a distinct commit.